Update the form_build_id if needed.
In some cases the form backend needs to update form_build_id after
modifying the form cache. The JS needs to modify the hidden input field
accordingly if this is the case.

Having this in the stripe module is not ideal (webform_paymethod_select)
would ideally take care of that perhaps by providing more API for the
AJAX callbacks. That’s not the case but this still has to be done somewhere.